Add missing tags file (#5)
With the current version, after installing this plugin as a submodule, 
a tag file is automatically generated when a file is opened in Vim.

However, it creates an annoying error when it keeps requiring to stage that 
tags file. And since it is a submodule, I cannot push that tags file to GitHub.
So I generate a tags file and add it to the current plugin so that no new tags 
file is generated which solves the problem.